+++ src/index.ts
@@ -3,6 +3,7 @@ import * as noAsyncpipe from "./rules/no-asyncpipe.js";
 import * as noChangedetectorref from "./rules/no-changedetectorref.js";
 import * as noContentDecorator from "./rules/no-content-decorator.js";
 import * as noDetectchangesTesting from "./rules/no-detectchanges-testing.js";
+import * as noDirectiveWritableProperty from "./rules/no-directive-writable-property.js";
 import * as noEagerChangeDetection from "./rules/no-eager-change-detection.js";
 import * as noInputDecorator from "./rules/no-input-decorator.js";
 import * as noNgaftercontentchecked from "./rules/no-ngaftercontentchecked.js";
@@ -19,7 +20,6 @@ import * as noOutputDecorator from "./rules/no-output-decorator.js";
 import * as noProvidezonechangedetection from "./rules/no-providezonechangedetection.js";
 import * as noSubscribeInComponentConstructor from "./rules/no-subscribe-in-component-constructor.js";
 import * as noViewDecorator from "./rules/no-view-decorator.js";
-import * as noWritableDirectiveProperty from "./rules/no-writable-directive-property.js";
 import * as noZonejsImport from "./rules/no-zonejs-import.js";
 import * as noZonejsTestingFunctions from "./rules/no-zonejs-testing-functions.js";
 
@@ -43,7 +43,7 @@ const plugin = {
   rules: {
     [noZonejsImport.ruleName]: noZonejsImport.ruleDefinition,
     [noProvidezonechangedetection.ruleName]: noProvidezonechangedetection.ruleDefinition,
-    [noWritableDirectiveProperty.ruleName]: noWritableDirectiveProperty.ruleDefinition,
+    [noDirectiveWritableProperty.ruleName]: noDirectiveWritableProperty.ruleDefinition,
     [noEagerChangeDetection.ruleName]: noEagerChangeDetection.ruleDefinition,
     [noNgoninit.ruleName]: noNgoninit.ruleDefinition,
     [noNgdocheck.ruleName]: noNgdocheck.ruleDefinition,
@@ -86,7 +86,7 @@ const recommended: ConfigObject = {
   },
   rules: {
     ...minimal.rules,
-    [`${name}/${noWritableDirectiveProperty.ruleName}`]: "error",
+    [`${name}/${noDirectiveWritableProperty.ruleName}`]: "error",
     [`${name}/${noEagerChangeDetection.ruleName}`]: "error",
     [`${name}/${noNgoninit.ruleName}`]: "error",
     [`${name}/${noNgdocheck.ruleName}`]: "error",