+++ src/rules/no-directive-writable-property.ts
@@ -2,8 +2,8 @@ import type { RuleDefinition } from "@eslint/core";
 import type { TSESTree } from "@typescript-eslint/utils";
 import { isInAngularClass } from "../utils/angular-class-decorator";
 
-export const ruleName = "no-writable-directive-property";
-const messageId = "noWritableDirectiveProperty";
+export const ruleName = "no-directive-writable-property";
+const messageId = "noDirectiveWritableProperty";
 
 export const ruleDefinition: RuleDefinition = {
   meta: {
@@ -13,7 +13,7 @@ export const ruleDefinition: RuleDefinition = {
     },
     docs: {
       description: `Checks that public and protected properties of components and directives are readonly.`,
-      url: 'https://github.com/cyrilletuzi/angular-eslint-zoneless/blob/main/docs/rules/NO_WRITABLE_DIRECTIVE_PROPERTY.md',
+      url: 'https://github.com/cyrilletuzi/angular-eslint-zoneless/blob/main/docs/rules/NO_DIRECTIVE_WRITABLE_PROPERTY.md',
       recommended: true,
     },
     schema: [],
@@ -21,7 +21,7 @@ export const ruleDefinition: RuleDefinition = {
   create(context) {
     return {
       /* React to: explicit `public` properties, explicit `protected` properties, implicit public properties which are not `#private` */
-      "PropertyDefinition[accessibility='public']:not([readonly=true]), PropertyDefinition[accessibility='protected']:not([readonly=true]), PropertyDefinition:not([accessibility]):not([readonly=true]):not([key.type='PrivateIdentifier'])"(node: TSESTree.MethodDefinition) {
+      "PropertyDefinition:not([accessibility='public']):not([readonly=true]), PropertyDefinition[accessibility='protected']:not([readonly=true]), PropertyDefinition:not([accessibility]):not([readonly=true]):not([key.type='PrivateIdentifier'])"(node: TSESTree.MethodDefinition) {
         if (isInAngularClass(node, ["Component", "Directive"])) {
           context.report({
             node,