+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.4.0 (2026-05-17)
 
-- feat: add `no-writable-directive-property` rule (in recommended preset)
+- feat: add `no-directive-writable-property` rule (in recommended preset)
 
 ## 1.3.0 (2026-05-11)